requirements.txt
# Core web framework
fastapi==0.114.0
uvicorn==0.37.0
httpx==0.28.1

# Environment variables
python-dotenv==1.1.1

# OpenAI SDK (official)
openai==1.99.1

# Audio processing
numpy==2.1.3
torch==2.8.0
torchaudio==2.8.0
torch-audiomentations==0.12.0

# Optional: used by torch hub (silero-vad)
soundfile==0.13.1

# Utilities
pydantic==2.11.9
pydantic[email]

# For WebSocket support
websockets==15.0.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Message Queue & Circuit Breaker
aio_pika==9.0.0
pybreaker==1.4.1
tenacity==9.1.2

# Rate Limiting
slowapi==0.1.9